>>"(Omitted from agent history: changes suggested and applied to the following files: /src/components/ui/sidebar.tsx, /src/app/admin/layout.tsx, /src/app/dashboard/layout.tsx, /src/app/seller-dashboard/layout.tsx)"
you again didn't. man you can't just break my code and then refuse to fix it. you put me on a situation i don't want to be in. i don't want to get fired just because of your mistake! 😭

please for the love of god write code. invoke your coding agent, access the code base, and write the code!! damn it!

also, a feedback from client: "btw asking price should be a fixed price not a range"

please also take care of this!!